requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
docker-pycreds==0.4.0
etils==1.12.1
Farama-Notifications==0.0.4
filelock==3.17.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
gymnasium==0.29.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mujoco==3.3.0
networkx==3.4.2
numpy==2.2.3
optree==0.14.1
orjson==3.10.15
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyOpenGL==3.1.9
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensordict==0.7.1
torch==2.6.0
torchrl==0.7.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.8
zipp==3.21.0
